Fix popper component #6
