+++ user-manager/src/components/common/Popper/index.tsx
@@ -1,13 +1,39 @@
 import '@components/common/Popper/Popper.css';
-import { IPopperProps } from '@types/interfaces';
-import usePopper from '@helpers/popper';
-import { useEffect } from 'react';
+import { useEffect, useRef, useState } from 'react';
 
-const Popper = ({ open = false, icon, children, options }: IPopperProps) => {
-  const { ref, isOpen, setIsOpen, handleTogglePopper } = usePopper();
+import useBackDrop from '@hooks/useBackDrop';
+
+interface IPopperProps {
+  open?: boolean;
+  icon?: string;
+  children?: string;
+  options: {
+    text?: string;
+    onClick(event: React.MouseEvent): void;
+  }[];
+}
+
+const Popper = ({
+  open = false,
+  icon,
+  children,
+  options
+}: IPopperProps) => {
+  const optionRef = useRef(null);
+  const [showOption, setShowOption] = useState(false);
+
+  const closeOption = () => {
+    setShowOption(false);
+  };
+
+  useBackDrop(optionRef, closeOption);
+
+  const handleTogglePopper = () => {
+    setShowOption(true);
+  };
 
   useEffect(() => {
-    setIsOpen(open);
+    setShowOption(open);
   }, [open]);
 
   return (
@@ -16,21 +42,19 @@ const Popper = ({ open = false, icon, children, options }: IPopperProps) => {
         className='btn-primary btn-popper'
         type='button'
         onClick={handleTogglePopper}
-        ref={ref as React.RefObject<HTMLButtonElement>}
-      >
+        ref={optionRef}>
         <img className='plus-icon' alt={`${children}-icon`} src={icon} />
         {children}
       </button>
 
-      {isOpen && (
+      {showOption && (
         <div className='popper-option'>
           {options.map((option, index) => (
             <button
               key={index}
               className='btn-popper btn-option'
               type='button'
-              onClick={option.onClick}
-            >
+              onClick={option.onClick}>
               {option.text}
             </button>
           ))}
